backend/requirements.txt
django==5.1.7
djangorestframework==3.15.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
openai==1.6.0
pillow==10.2.0
django-storages==1.14.2
requests==2.31.0
celery==5.3.6
redis==5.0.1
